Display raw particle concentration on device when no test is running
This is potentially useful when debugging issues with the device, and/or
for testers to verify that in-specimen particle levels have settled.

Application such as Incolata can/do display the same data on screen
already, but showing it on the device too is still nice (in particular
for other client applications/use cases).