test: clear display on start
Because: we might be showing non-test content (e.g. at time of writing, the
device lib will render the raw concentrations on screen, but we stop updating
those during the test - and even if we want to show them during the test,
the test implementation should be responsible for updates during the test).