test: repeatedly send valve switch until confirmation received
Sometimes the very first valve switch is silently ignored - often along with the
following ClearDisplay (my 8020A simply ignores these messages without returning
an error). This in turn results in the test remaining stuck - we keep waiting
for a ValveAmbient response which never arrives because the device isn't
switching. It's not clear why the ValveAmbient message is being ignored, but
my testing suggests that sending a second ValveAmbient (VN) will get the message
across.

This has never happened during a test. The difference might be as simple as
follows:

- During a test, valve messages are sent immediately after receiving a sample.
- When starting a test, the initial ValveAmbient might be sent while the device
  is sending the sample - and gets swallowed?

(This is just one hypothesis of many.)